fix: resolve CodeRabbit feedback (2 issues)
- De-duplicate retry logic into shared utils/retry.py module
- Redact Claude response payloads from exception messages to avoid PII exposure
