+++ src/steps/utils/__init__.py
@@ -1,0 +1,1 @@
+# Utils package for shared helper modules
